requirements.txt
Keras >= 2.0.0
tensorflow-gpu >= 1.0.0